refactor: streamline scene configuration loading and validation
