+++ custom_components/stateful_scenes/__init__.py
@@ -5,6 +5,9 @@ from __future__ import annotations
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import Platform
 from homeassistant.core import HomeAssistant
+import aiofiles
+import yaml
+import os
 
 from .StatefulScenes import Hub, Scene
 from .const import (
@@ -12,6 +15,8 @@ from .const import (
     CONF_SCENE_PATH,
     CONF_NUMBER_TOLERANCE,
     CONF_ENABLE_DISCOVERY,
+    StatefulScenesYamlNotFound,
+    StatefulScenesYamlInvalid,
 )
 from .discovery import DiscoveryManager
 
@@ -26,12 +31,19 @@ async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     """Set up this integration using UI."""
     hass.data.setdefault(DOMAIN, {})
     is_hub = entry.data.get("hub", None)
+
     if is_hub is None:
         is_hub = CONF_SCENE_PATH in entry.data
+
     if is_hub:
+        if entry.data.get(CONF_SCENE_PATH, None) is None:
+            raise StatefulScenesYamlNotFound("Scenes file not specified.")
+
+        scene_confs = await load_scenes_file(entry.data[CONF_SCENE_PATH])
+
         hass.data[DOMAIN][entry.entry_id] = Hub(
             hass=hass,
-            scene_path=entry.data[CONF_SCENE_PATH],
+            scene_confs=scene_confs,
             number_tolerance=entry.data[CONF_NUMBER_TOLERANCE],
         )
 
@@ -58,3 +70,23 @@ async def async_reload_entry(hass: HomeAssistant, entry: ConfigEntry) -> None:
     """Reload config entry."""
     await async_unload_entry(hass, entry)
     await async_setup_entry(hass, entry)
+
+
+async def load_scenes_file(scene_path) -> list:
+    """Load scenes from yaml file."""
+    # check if file exists
+    if scene_path is None:
+        raise StatefulScenesYamlNotFound("Scenes file not specified.")
+    if not os.path.exists(scene_path):
+        raise StatefulScenesYamlNotFound("No scenes file " + scene_path)
+
+    try:
+        async with aiofiles.open(scene_path, encoding="utf-8") as f:
+            scenes_confs = yaml.load(await f.read(), Loader=yaml.FullLoader)
+    except OSError as err:
+        raise StatefulScenesYamlInvalid("No scenes found in " + scene_path) from err
+
+    if not scenes_confs or not isinstance(scenes_confs, list):
+        raise StatefulScenesYamlInvalid("No scenes found in " + scene_path)
+
+    return scenes_confs
